requirements.txt
fastapi
uvicorn
pyserial


# install the requirements using the following command:-
'''pip install -r requirements.txt'''
